requirements.txt
psycopg[binary]
python-dotenv
requests
sqlmodel
